backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.24.1
pydantic==2.9.2
pydantic-settings==2.5.2

SQLAlchemy==2.0.35
psycopg2-binary
asyncpg==0.29.0

python-jose[cryptography]==3.3.0
python-multipart>=0.0.6
web3==6.20.3
eth-account==0.11.3